Revert "Refactor building the Query struct (#5893)"
This reverts commit 7a11f5ec408126283544a2de83f30940479f4758.
