+++ lib/plausible/stats/filters/query_parser.ex
@@ -1,9 +1,30 @@
-defmodule Plausible.Stats.QueryParser do
+defmodule Plausible.Stats.Filters.QueryParser do
   @moduledoc false
 
   use Plausible
 
-  alias Plausible.Stats.{Filters, Metrics, DateTimeRange, JSONSchema}
+  alias Plausible.Stats.{TableDecider, Filters, Metrics, DateTimeRange, JSONSchema, Time}
+
+  @default_include %{
+    imports: false,
+    # `include.imports_meta` can be true even when `include.imports`
+    # is false. Even if we don't want to include imported data, we
+    # might still want to know whether imported data can be toggled
+    # on/off on the dashboard.
+    imports_meta: false,
+    time_labels: false,
+    total_rows: false,
+    trim_relative_date_range: false,
+    comparisons: nil,
+    legacy_time_on_page_cutoff: nil
+  }
+
+  @default_pagination %{
+    limit: 10_000,
+    offset: 0
+  }
+
+  def default_include(), do: @default_include
 
   def parse(site, schema_type, params, now \\ nil) when is_map(params) do
     now = now || Plausible.Stats.Query.Test.get_fixed_now()
@@ -14,24 +35,56 @@ defmodule Plausible.Stats.QueryParser do
          {:ok, raw_time_range} <-
            parse_time_range(site, Map.get(params, "date_range"), date, now),
          utc_time_range = raw_time_range |> DateTimeRange.to_timezone("Etc/UTC"),
-         {:ok, metrics} <- parse_metrics(Map.fetch!(params, "metrics")),
-         {:ok, filters} <- parse_filters(params["filters"]),
-         {:ok, dimensions} <- parse_dimensions(params["dimensions"]),
-         {:ok, order_by} <- parse_order_by(params["order_by"]),
-         {:ok, pagination} <- parse_pagination(params["pagination"]),
-         {:ok, include} <- parse_include(params["include"], site) do
-      {:ok,
-       Plausible.Stats.ParsedQueryParams.new!(%{
-         now: now,
-         utc_time_range: utc_time_range,
-         metrics: metrics,
-         filters: filters,
-         dimensions: dimensions,
-         order_by: order_by,
-         pagination: pagination,
-         include: include
-       })}
+         {:ok, metrics} <- parse_metrics(Map.get(params, "metrics", [])),
+         {:ok, filters} <- parse_filters(Map.get(params, "filters", [])),
+         {:ok, preloaded_segments} <-
+           Plausible.Segments.Filters.preload_needed_segments(site, filters),
+         {:ok, filters} <-
+           Plausible.Segments.Filters.resolve_segments(filters, preloaded_segments),
+         {:ok, dimensions} <- parse_dimensions(Map.get(params, "dimensions", [])),
+         {:ok, order_by} <- parse_order_by(Map.get(params, "order_by")),
+         {:ok, include} <- parse_include(Map.get(params, "include", %{}), site),
+         {:ok, pagination} <- parse_pagination(Map.get(params, "pagination", %{})),
+         {preloaded_goals, revenue_warning, revenue_currencies} <-
+           preload_goals_and_revenue(site, metrics, filters, dimensions),
+         consolidated_site_ids = get_consolidated_site_ids(site),
+         query = %{
+           now: now,
+           consolidated_site_ids: consolidated_site_ids,
+           input_date_range: Map.get(params, "date_range"),
+           metrics: metrics,
+           filters: filters,
+           utc_time_range: utc_time_range,
+           dimensions: dimensions,
+           order_by: order_by,
+           timezone: site.timezone,
+           include: include,
+           pagination: pagination,
+           preloaded_goals: preloaded_goals,
+           revenue_warning: revenue_warning,
+           revenue_currencies: revenue_currencies
+         },
+         :ok <- validate_order_by(query),
+         :ok <- validate_custom_props_access(site, query),
+         :ok <- validate_toplevel_only_filter_dimension(query),
+         :ok <- validate_special_metrics_filters(query),
+         :ok <- validate_behavioral_filters(query),
+         :ok <- validate_filtered_goals_exist(query),
+         :ok <- validate_revenue_metrics_access(site, query),
+         :ok <- validate_metrics(query),
+         :ok <- validate_include(query) do
+      {:ok, query}
     end
+  end
+
+  on_ee do
+    def get_consolidated_site_ids(%Plausible.Site{} = site) do
+      if Plausible.Sites.consolidated?(site) do
+        Plausible.ConsolidatedView.Cache.get(site.domain)
+      end
+    end
+  else
+    def get_consolidated_site_ids(_site), do: nil
   end
 
   def parse_date_range_pair(site, [from, to]) when is_binary(from) and is_binary(to) do
@@ -57,7 +110,7 @@ defmodule Plausible.Stats.QueryParser do
     parse_list(filters, &parse_filter/1)
   end
 
-  def parse_filters(nil), do: {:ok, nil}
+  def parse_filters(_invalid_metrics), do: {:error, "Invalid filters passed."}
 
   defp parse_filter(filter) do
     with {:ok, operator} <- parse_operator(filter),
@@ -255,8 +308,6 @@ defmodule Plausible.Stats.QueryParser do
     )
   end
 
-  defp parse_dimensions(nil), do: {:ok, nil}
-
   def parse_order_by(order_by) when is_list(order_by) do
     parse_list(order_by, &parse_order_by_entry/1)
   end
@@ -308,19 +359,16 @@ defmodule Plausible.Stats.QueryParser do
   defp parse_order_direction(entry), do: {:error, "Invalid order_by entry '#{i(entry)}'."}
 
   def parse_include(include, site) when is_map(include) do
-    with {:ok, include} <- atomize_include_keys(include) do
-      update_comparisons_date_range(include, site)
+    with {:ok, include} <- atomize_include_keys(include),
+         {:ok, include} <- update_comparisons_date_range(include, site) do
+      {:ok, Map.merge(@default_include, include)}
     end
   end
 
-  def parse_include(nil, _site), do: {:ok, nil}
   def parse_include(include, _site), do: {:error, "Invalid include '#{i(include)}'."}
 
   defp atomize_include_keys(map) do
-    expected_keys =
-      Plausible.Stats.ParsedQueryParams.default_include()
-      |> Map.keys()
-      |> Enum.map(&Atom.to_string/1)
+    expected_keys = @default_include |> Map.keys() |> Enum.map(&Atom.to_string/1)
 
     if Map.keys(map) |> Enum.all?(&(&1 in expected_keys)) do
       {:ok, atomize_keys(map)}
@@ -338,11 +386,8 @@ defmodule Plausible.Stats.QueryParser do
   defp update_comparisons_date_range(include, _site), do: {:ok, include}
 
   defp parse_pagination(pagination) when is_map(pagination) do
-    {:ok,
-     Map.merge(Plausible.Stats.ParsedQueryParams.default_pagination(), atomize_keys(pagination))}
+    {:ok, Map.merge(@default_pagination, atomize_keys(pagination))}
   end
-
-  defp parse_pagination(nil), do: {:ok, nil}
 
   defp atomize_keys(map) when is_map(map) do
     Map.new(map, fn {key, value} ->
@@ -384,12 +429,273 @@ defmodule Plausible.Stats.QueryParser do
     end
   end
 
+  defp validate_order_by(query) do
+    if query.order_by do
+      valid_values = query.metrics ++ query.dimensions
+
+      invalid_entry =
+        Enum.find(query.order_by, fn {value, _direction} ->
+          not Enum.member?(valid_values, value)
+        end)
+
+      case invalid_entry do
+        nil ->
+          :ok
+
+        _ ->
+          {:error,
+           "Invalid order_by entry '#{i(invalid_entry)}'. Entry is not a queried metric or dimension."}
+      end
+    else
+      :ok
+    end
+  end
+
+  def preload_goals_and_revenue(site, metrics, filters, dimensions) do
+    preloaded_goals =
+      Plausible.Stats.Goals.preload_needed_goals(site, dimensions, filters)
+
+    {revenue_warning, revenue_currencies} =
+      preload_revenue(site, preloaded_goals, metrics, dimensions)
+
+    {
+      preloaded_goals,
+      revenue_warning,
+      revenue_currencies
+    }
+  end
+
+  @only_toplevel ["event:goal", "event:hostname"]
+  defp validate_toplevel_only_filter_dimension(query) do
+    not_toplevel =
+      query.filters
+      |> Filters.dimensions_used_in_filters(min_depth: 1, behavioral_filters: :ignore)
+      |> Enum.filter(&(&1 in @only_toplevel))
+
+    if Enum.count(not_toplevel) > 0 do
+      {:error,
+       "Invalid filters. Dimension `#{List.first(not_toplevel)}` can only be filtered at the top level."}
+    else
+      :ok
+    end
+  end
+
+  @special_metrics [:conversion_rate, :group_conversion_rate]
+  defp validate_special_metrics_filters(query) do
+    special_metric? = Enum.any?(@special_metrics, &(&1 in query.metrics))
+
+    deep_custom_property? =
+      query.filters
+      |> Filters.dimensions_used_in_filters(min_depth: 1)
+      |> Enum.any?(fn dimension -> String.starts_with?(dimension, "event:props:") end)
+
+    if special_metric? and deep_custom_property? do
+      {:error,
+       "Invalid filters. When `conversion_rate` or `group_conversion_rate` metrics are used, custom property filters can only be used on top level."}
+    else
+      :ok
+    end
+  end
+
+  defp validate_behavioral_filters(query) do
+    query.filters
+    |> Filters.traverse(0, fn behavioral_depth, operator ->
+      if operator in [:has_done, :has_not_done] do
+        behavioral_depth + 1
+      else
+        behavioral_depth
+      end
+    end)
+    |> Enum.reduce_while(:ok, fn {[_operator, dimension | _rest], behavioral_depth}, :ok ->
+      cond do
+        behavioral_depth == 0 ->
+          # ignore non-behavioral filters
+          {:cont, :ok}
+
+        behavioral_depth > 1 ->
+          {:halt,
+           {:error,
+            "Invalid filters. Behavioral filters (has_done, has_not_done) cannot be nested."}}
+
+        not String.starts_with?(dimension, "event:") ->
+          {:halt,
+           {:error,
+            "Invalid filters. Behavioral filters (has_done, has_not_done) can only be used with event dimension filters."}}
+
+        true ->
+          {:cont, :ok}
+      end
+    end)
+  end
+
+  defp validate_filtered_goals_exist(query) do
+    # Note: We don't check :contains goal filters since it's acceptable if they match nothing.
+    goal_filter_clauses =
+      query.filters
+      |> Filters.all_leaf_filters()
+      |> Enum.flat_map(fn
+        [:is, "event:goal", clauses] -> clauses
+        _ -> []
+      end)
+
+    if length(goal_filter_clauses) > 0 do
+      configured_goal_names =
+        query.preloaded_goals.all
+        |> Enum.map(&Plausible.Goal.display_name/1)
+
+      validate_list(goal_filter_clauses, &validate_goal_filter(&1, configured_goal_names))
+    else
+      :ok
+    end
+  end
+
+  on_ee do
+    alias Plausible.Stats.Goal.Revenue
+
+    def preload_revenue(site, preloaded_goals, metrics, dimensions) do
+      Revenue.preload(site, preloaded_goals, metrics, dimensions)
+    end
+
+    defp validate_revenue_metrics_access(site, query) do
+      if Revenue.requested?(query.metrics) and not Revenue.available?(site) do
+        {:error, "The owner of this site does not have access to the revenue metrics feature."}
+      else
+        :ok
+      end
+    end
+  else
+    defp preload_revenue(_site, _preloaded_goals, _metrics, _dimensions), do: {nil, %{}}
+
+    defp validate_revenue_metrics_access(_site, _query), do: :ok
+  end
+
+  defp validate_goal_filter(clause, configured_goal_names) do
+    if Enum.member?(configured_goal_names, clause) do
+      :ok
+    else
+      {:error,
+       "Invalid filters. The goal `#{clause}` is not configured for this site. Find out how to configure goals here: https://plausible.io/docs/stats-api#filtering-by-goals"}
+    end
+  end
+
+  defp validate_custom_props_access(site, query) do
+    allowed_props = Plausible.Props.allowed_for(site, bypass_setup?: true)
+
+    validate_custom_props_access(site, query, allowed_props)
+  end
+
+  defp validate_custom_props_access(_site, _query, :all), do: :ok
+
+  defp validate_custom_props_access(_site, query, allowed_props) do
+    valid? =
+      query.filters
+      |> Filters.dimensions_used_in_filters()
+      |> Enum.concat(query.dimensions)
+      |> Enum.all?(fn
+        "event:props:" <> prop -> prop in allowed_props
+        _ -> true
+      end)
+
+    if valid? do
+      :ok
+    else
+      {:error, "The owner of this site does not have access to the custom properties feature."}
+    end
+  end
+
+  defp validate_metrics(query) do
+    with :ok <- validate_list(query.metrics, &validate_metric(&1, query)) do
+      TableDecider.validate_no_metrics_dimensions_conflict(query)
+    end
+  end
+
+  defp validate_metric(metric, query) when metric in [:conversion_rate, :group_conversion_rate] do
+    if Enum.member?(query.dimensions, "event:goal") or
+         Filters.filtering_on_dimension?(query, "event:goal", behavioral_filters: :ignore) do
+      :ok
+    else
+      {:error, "Metric `#{metric}` can only be queried with event:goal filters or dimensions."}
+    end
+  end
+
+  defp validate_metric(:scroll_depth = metric, query) do
+    page_dimension? = Enum.member?(query.dimensions, "event:page")
+    toplevel_page_filter? = not is_nil(Filters.get_toplevel_filter(query, "event:page"))
+
+    if page_dimension? or toplevel_page_filter? do
+      :ok
+    else
+      {:error, "Metric `#{metric}` can only be queried with event:page filters or dimensions."}
+    end
+  end
+
+  defp validate_metric(:exit_rate = metric, query) do
+    case {query.dimensions, TableDecider.sessions_join_events?(query)} do
+      {["visit:exit_page"], false} ->
+        :ok
+
+      {["visit:exit_page"], true} ->
+        {:error, "Metric `#{metric}` cannot be queried when filtering on event dimensions."}
+
+      _ ->
+        {:error,
+         "Metric `#{metric}` requires a `\"visit:exit_page\"` dimension. No other dimensions are allowed."}
+    end
+  end
+
+  defp validate_metric(:views_per_visit = metric, query) do
+    cond do
+      Filters.filtering_on_dimension?(query, "event:page", behavioral_filters: :ignore) ->
+        {:error, "Metric `#{metric}` cannot be queried with a filter on `event:page`."}
+
+      length(query.dimensions) > 0 ->
+        {:error, "Metric `#{metric}` cannot be queried with `dimensions`."}
+
+      true ->
+        :ok
+    end
+  end
+
+  defp validate_metric(:time_on_page = metric, query) do
+    cond do
+      Enum.member?(query.dimensions, "event:page") ->
+        :ok
+
+      Filters.filtering_on_dimension?(query, "event:page", behavioral_filters: :ignore) ->
+        :ok
+
+      true ->
+        {:error, "Metric `#{metric}` can only be queried with event:page filters or dimensions."}
+    end
+  end
+
+  defp validate_metric(_, _), do: :ok
+
+  defp validate_include(query) do
+    time_dimension? = Enum.any?(query.dimensions, &Time.time_dimension?/1)
+
+    if query.include.time_labels and not time_dimension? do
+      {:error, "Invalid include.time_labels: requires a time dimension."}
+    else
+      :ok
+    end
+  end
+
   defp i(value), do: inspect(value, charlists: :as_lists)
 
   defp parse_list(list, parser_function) do
     Enum.reduce_while(list, {:ok, []}, fn value, {:ok, results} ->
       case parser_function.(value) do
         {:ok, result} -> {:cont, {:ok, results ++ [result]}}
+        {:error, _} = error -> {:halt, error}
+      end
+    end)
+  end
+
+  defp validate_list(list, parser_function) do
+    Enum.reduce_while(list, :ok, fn value, :ok ->
+      case parser_function.(value) do
+        :ok -> {:cont, :ok}
         {:error, _} = error -> {:halt, error}
       end
     end)
