+++ lib/plausible/stats/query.ex
@@ -18,7 +18,7 @@ defmodule Plausible.Stats.Query do
             timezone: nil,
             legacy_breakdown: false,
             preloaded_goals: [],
-            include: Plausible.Stats.ParsedQueryParams.default_include(),
+            include: Plausible.Stats.Filters.QueryParser.default_include(),
             debug_metadata: %{},
             pagination: nil,
             # Revenue metric specific metadata
@@ -34,40 +34,45 @@ defmodule Plausible.Stats.Query do
             smear_session_metrics: false
 
   require OpenTelemetry.Tracer, as: Tracer
-
-  alias Plausible.Stats.{
-    DateTimeRange,
-    Imported,
-    Legacy,
-    Comparisons,
-    QueryParser,
-    ParsedQueryParams,
-    QueryBuilder
-  }
+  alias Plausible.Stats.{DateTimeRange, Filters, Imported, Legacy, Comparisons}
 
   @type t :: %__MODULE__{}
 
-  def parse_and_build(
+  def build(
         %Plausible.Site{domain: domain} = site,
         schema_type,
         %{"site_id" => domain} = params,
         debug_metadata \\ %{}
       ) do
-    with {:ok, %ParsedQueryParams{} = parsed_query_params} <-
-           QueryParser.parse(site, schema_type, params) do
-      QueryBuilder.build(site, parsed_query_params, params, debug_metadata)
+    with {:ok, query_data} <- Filters.QueryParser.parse(site, schema_type, params) do
+      query =
+        %__MODULE__{
+          debug_metadata: debug_metadata,
+          site_id: site.id,
+          site_native_stats_start_at: site.native_stats_start_at
+        }
+        |> struct!(Map.to_list(query_data))
+        |> set_time_on_page_data(site)
+        |> put_comparison_utc_time_range()
+        |> put_imported_opts(site)
+
+      on_ee do
+        query = Plausible.Stats.Sampling.put_threshold(query, site, params)
+      end
+
+      {:ok, query}
     end
   end
 
-  def parse_and_build!(site, schema_type, params, debug_metadata \\ %{}) do
-    case parse_and_build(site, schema_type, params, debug_metadata) do
+  def build!(site, schema_type, params, debug_metadata \\ %{}) do
+    case build(site, schema_type, params, debug_metadata) do
       {:ok, query} -> query
       {:error, reason} -> raise "Failed to build query: #{inspect(reason)}"
     end
   end
 
   @doc """
-  Builds query from old-style stats APIv1 params. New code should use `Query.parse_and_build`.
+  Builds query from old-style stats APIv1 params. New code should use `Query.build`.
   """
   def from(site, params, debug_metadata \\ %{}, now \\ nil) do
     Legacy.QueryBuilder.from(site, params, debug_metadata, now)
@@ -138,6 +143,13 @@ defmodule Plausible.Stats.Query do
     put_imported_opts(query, nil)
   end
 
+  def put_comparison_utc_time_range(%__MODULE__{include: %{comparisons: nil}} = query), do: query
+
+  def put_comparison_utc_time_range(%__MODULE__{include: %{comparisons: comparison_opts}} = query) do
+    datetime_range = Comparisons.get_comparison_utc_time_range(query, comparison_opts)
+    struct!(query, comparison_utc_time_range: datetime_range)
+  end
+
   def put_imported_opts(query, site) do
     requested? = query.include.imports
 
@@ -176,6 +188,15 @@ defmodule Plausible.Stats.Query do
       end
 
     in_comparison_range ++ in_range
+  end
+
+  def set_time_on_page_data(query, site) do
+    struct!(query,
+      time_on_page_data: %{
+        new_metric_visible: Plausible.Stats.TimeOnPage.new_time_on_page_visible?(site),
+        cutoff_date: site.legacy_time_on_page_cutoff
+      }
+    )
   end
 
   @spec get_skip_imported_reason(t()) ::
