+++ lib/plausible_web/live/goal_settings/form.ex
@@ -545,7 +545,7 @@ defmodule PlausibleWeb.Live.GoalSettings.Form do
 
   def suggest_page_paths(input, site) do
     query =
-      Plausible.Stats.Query.parse_and_build!(
+      Plausible.Stats.Query.build!(
         site,
         :internal,
         %{
