+++ lib/plausible_web/live/shields/hostname_rules.ex
@@ -252,7 +252,7 @@ defmodule PlausibleWeb.Live.Shields.HostnameRules do
 
   def suggest_hostnames(input, _options, site) do
     query =
-      Plausible.Stats.Query.parse_and_build!(
+      Plausible.Stats.Query.build!(
         site,
         :internal,
         %{
