+++ test/plausible/stats/query_optimizer_test.exs
@@ -1,7 +1,7 @@
-defmodule Plausible.Stats.Query.QueryOptimizerTest do
+defmodule Plausible.Stats.QueryOptimizerTest do
   use Plausible.DataCase, async: true
 
-  alias Plausible.Stats.{Query, QueryOptimizer, DateTimeRange, ParsedQueryParams}
+  alias Plausible.Stats.{Query, QueryOptimizer, DateTimeRange}
 
   @default_params %{metrics: [:visitors]}
 
@@ -154,6 +154,8 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
   end
 
   describe "trim_relative_date_range" do
+    alias Plausible.Stats.Filters.QueryParser
+
     test "trims current month period when flag is set" do
       now = DateTime.new!(~D[2024-01-15], ~T[12:00:00], "UTC")
 
@@ -163,7 +165,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "month",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range.first == ~U[2024-01-01 00:00:00Z]
@@ -179,7 +181,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "year",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range.first == ~U[2024-01-01 00:00:00Z]
@@ -195,7 +197,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "day",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range.first == ~U[2024-01-15 00:00:00Z]
@@ -212,7 +214,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "month",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range == original_range
@@ -228,7 +230,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "year",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range == original_range
@@ -244,7 +246,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "day",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range == original_range
@@ -262,7 +264,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           timezone: "UTC",
           include:
             Map.merge(
-              ParsedQueryParams.default_include(),
+              QueryParser.default_include(),
               %{comparisons: %{mode: "previous_period"}, trim_relative_date_range: true}
             )
         })
@@ -281,7 +283,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           timezone: "UTC",
           include:
             Map.merge(
-              ParsedQueryParams.default_include(),
+              QueryParser.default_include(),
               %{comparisons: %{mode: "previous_period"}, trim_relative_date_range: true}
             )
         })
@@ -300,7 +302,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "month",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, false)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, false)
         })
 
       assert result.utc_time_range == original_range
@@ -316,7 +318,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "month",
           now: now,
           timezone: "UTC",
-          include: ParsedQueryParams.default_include()
+          include: QueryParser.default_include()
         })
 
       assert result.utc_time_range == original_range
@@ -333,7 +335,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "7d",
           now: now,
           timezone: "UTC",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       assert result.utc_time_range == original_range
@@ -350,7 +352,7 @@ defmodule Plausible.Stats.Query.QueryOptimizerTest do
           input_date_range: "year",
           now: now,
           timezone: "America/New_York",
-          include: Map.put(ParsedQueryParams.default_include(), :trim_relative_date_range, true)
+          include: Map.put(QueryParser.default_include(), :trim_relative_date_range, true)
         })
 
       nyc_mar_15_end =
