+++ test/plausible/stats/query_test.exs
@@ -1,8 +1,9 @@
-defmodule Plausible.Stats.Query.QueryFromTest do
+defmodule Plausible.Stats.QueryTest do
   use Plausible.DataCase, async: true
   use Plausible.Teams.Test
   alias Plausible.Stats.Query
   alias Plausible.Stats.Legacy.QueryBuilder
+  alias Plausible.Stats.Filters.QueryParser
   alias Plausible.Stats.DateTimeRange
 
   doctest Plausible.Stats.Legacy.QueryBuilder
